api/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
weasyprint==62.3
PyYAML==6.0.2
Jinja2==3.1.4
python-docx==1.1.2
rapidfuzz==3.9.6
unidecode==1.3.8
openai==1.3.0
requests==2.31.0
Pillow==10.1.0

# Firebase for file storage
firebase-admin==6.4.0
google-cloud-storage==2.14.0

# Gemini AI for manuscript analysis
google-generativeai==0.3.2

# Document processing libraries
mammoth==1.6.0
odfpy==1.4.1
striprtf==0.0.26
markdown==3.5.1
beautifulsoup4==4.12.2
ebooklib==0.18

# EPUB generation
pypandoc==1.13

# Additional utilities
lxml==5.0.0
pypdf2==3.0.1
